Reduce some OSGi code duplication with core Activator.
